[Feature/38]: 카카오 소셜로그인 환경변수 설정 및 응답 확인
