+++ src/main/java/juby/invest/global/security/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package juby.invest.domain.member.dto;
+package juby.invest.global.security.dto;
 
 import org.springframework.stereotype.Component;
