+++ src/main/java/juby/invest/global/security/filter/JwtAuthenticationFilter.java
@@ -1,9 +1,10 @@
-package juby.invest.global.jwt;
+package juby.invest.global.security.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import juby.invest.global.security.util.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
