+++ src/main/resources/application.yaml
@@ -40,6 +40,17 @@ spring:
             scope:
               - email
               - profile
+          kakao:
+            client-name: kakao
+            client-id: ${KAKAO_CLIENT_ID}
+            client-secret: ${KAKAO_CLIENT_SECRET}
+            client-authentication-method: client_secret_post
+            redirect-uri: http://localhost:8080/login/oauth2/code/kakao
+            authorization-grant-type: authorization_code
+            scope:
+              - profile_nickname
+#              - profile_image
+              - account_email
 
         provider:
           naver:
@@ -47,6 +58,12 @@ spring:
             token-uri: https://nid.naver.com/oauth2.0/token # Authorization Server 토큰 교환 주소
             user-info-uri: https://openapi.naver.com/v1/nid/me #Resource Server
             user-name-attribute: response
+
+          kakao:
+            authorization-uri: https://kauth.kakao.com/oauth/authorize
+            token-uri: https://kauth.kakao.com/oauth/token
+            user-info-uri: https://kapi.kakao.com/v2/user/me
+            user-name-attribute: id
 
 kis:
   mock:
@@ -70,4 +87,8 @@ jwt:
   refresh-token-validity: 1209600000 # 14일
 
 pinecone:
-    app-key: ${PINECONE_APPKEY}
+  app-key: ${PINECONE_APPKEY}
+
+logging:
+  level:
+    org.springframework.security: DEBUG
