+++ src/main/resources/templates/index.html
@@ -15,6 +15,6 @@
     <a href="/oauth2/authorization/google">구글 로그인</a>
 
     <button>카카오 로그인</button>
-    <a href="/oauth/authorization/kakao">카카오 로그인</a>
+    <a href="/oauth2/authorization/kakao">카카오 로그인</a>
 </body>
 </html>
